Refactor signing and checksum in .opnet binary build
Refactored the buildOpnetBinary function to compute and embed the checksum in the plugin metadata, and to accept a signing function for generating the signature. Updated CompileCommand and SignCommand to use the new signing flow and to log the checksum. Also updated the default opnetVersion in manifest creation and init to '>=0.0.1'.